requirements.txt
pyyaml
sympy
pytest
llama-cpp-python
requests
sentence-transformers
faiss-cpu
httpx
transformers
datasets
peft
accelerate
bitsandbytes
